+++ python/observation-publisher/publisher/__main__.py
@@ -3,7 +3,7 @@ from omegaconf import OmegaConf
 
 from publisher.config import PublisherConfig
 from publisher.observability_backend import new_observation_sink
-from publisher.prediction_log_consumer import new_consumer
+from publisher.observation_log_consumer import new_consumer
 
 
 @hydra.main(version_base=None, config_path="../conf", config_name="config")
@@ -11,7 +11,7 @@ def start_consumer(cfg: PublisherConfig) -> None:
     missing_keys: set[str] = OmegaConf.missing_keys(cfg)
     if missing_keys:
         raise RuntimeError(f"Got missing keys in config:\n{missing_keys}")
-    prediction_log_consumer = new_consumer(cfg.environment.consumer)
+    prediction_log_consumer = new_consumer(cfg.environment.observation_source)
     observation_sink = new_observation_sink(
         cfg.environment.observability_backend, cfg.environment.model
     )
